Float sens prediction vals

+++ python/interpret-core/interpret/blackbox/sensitivity.py
@@ -67,7 +67,7 @@ class MorrisSensitivity(ExplainerMixin):
 
         samples = self.sampler.sample()
         problem = self.sampler.gen_problem_from_data(self.data, self.feature_names)
-        analysis = morris.analyze(problem, samples, self.predict_fn(samples))
+        analysis = morris.analyze(problem, samples, self.predict_fn(samples).astype(float))
 
         mu = analysis["mu"]
         mu_star = analysis["mu_star"]
